feat(routes): /annotation-campaign/$campaignID/info
