--- frontend/src/view/annotation-campaign/[campaignID]/index.tsx
@@ -1,2 +1,0 @@
-export { AnnotationCampaignInfo } from './InfoTab'
-
