--- frontend/src/view/annotation-campaign/[campaignID]/styles.module.scss
@@ -1,56 +1,0 @@
-.tabs {
-  border-bottom: 1px solid rgba(var(--ion-color-medium-rgb), 0.25);
-  display: flex;
-  gap: 1rem;
-  padding-bottom: 0.5rem;
-
-  ion-button {
-    --padding-bottom: 0.25rem;
-    --padding-top: 0.5rem;
-    --padding-start: 1rem;
-    --padding-end: 1rem;
-    --border-radius: 0rem;
-    border-top-left-radius: 1rem;
-    border-top-right-radius: 1rem;
-    overflow: hidden;
-    border: 1px solid rgba(var(--ion-color-medium-rgb), 0);
-    border-bottom: none;
-
-    &.active {
-      border-color: rgba(var(--ion-color-medium-rgb), 0.5);
-    }
-  }
-}
-
-.info {
-  display: grid;
-  gap: 3rem;
-  align-content: start;
-  justify-content: start;
-  justify-items: start;
-  font-size: 0.9rem;
-  overflow: auto;
-  height: 100%;
-
-  .bloc {
-    display: grid;
-    gap: 0.5rem;
-    width: 100%;
-    justify-items: start;
-
-    > div {
-      width: 100%;
-      overflow: hidden;
-    }
-
-    .progress {
-      width: 20rem;
-    }
-  }
-
-  p {
-    word-break: break-word;
-    text-wrap: wrap;
-    width: 100%;
-  }
-}
